requirements.txt
appdirs==1.4.3
awsebcli==3.10.0
backports.ssl-match-hostname==3.5.0.1
blessed==1.14.1
botocore==1.5.26
cement==2.8.2
colorama==0.3.7
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.9.5
jmespath==0.9.2
MarkupSafe==1.0
numpy==1.12.1
packaging==16.8
pathspec==0.5.0
pyparsing==2.2.0
python-dateutil==2.6.0
PyYAML==3.12
requests==2.9.1
semantic-version==2.5.0
six==1.10.0
tabulate==0.7.5
termcolor==1.1.0
wcwidth==0.1.7
websocket-client==0.40.0
Werkzeug==0.12.1
biopython==1.68
